refactor pages and components
