+++ src/frontend/my-app/src/App.js
@@ -3,18 +3,18 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import Navbar from './Navbar'
 import HomePage from './HomePage';
 import CollabPage from './CollabPage';
-import CodeEditorPage from './components/CodeEditorPage';
+import CodeEditorPage from './pages/CodeEditorPage';
 
 function App() {
   return (
     <>
       <BrowserRouter>
         <Navbar />
-        <Container>
+        <Container >
           <Routes>
             <Route path="/" exact element={<HomePage />} />
             <Route path="/collab" exact element={<CollabPage />} />
-            <Route path="/question" exact element={<CodeEditorPage />} />
+            <Route path="/question/:questionId" element={<CodeEditorPage/>}/>
           </Routes>
         </Container>
       </BrowserRouter>
